+++ Libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: nriviere <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/11 16:42:35 by nriviere          #+#    #+#             */
-/*   Updated: 2022/11/11 16:47:42 by nriviere         ###   ########.fr       */
+/*   Updated: 2022/11/20 12:19:25 by nriviere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] && s2[i] && i < n && s1[i] == s2[i])
